+++ CHANGELOG.md
@@ -11,6 +11,8 @@ From v2.1.0.a2 to v2.2.0
 - The jupyter notebooks now pass a random state to ensure consistent results.
 - The default parameter values for `STFullyConnected` have changed from `n_epochs` = 1000 to `n_epochs` = 100, from `neurons_h1` = 4000 to `neurons_h1` = 256 and `neurons_hx` = 1000 to `neurons_hx` = 128.
 - Rename `HyperParameterOptimization` to `HyperparameterOptimization`.
+- Make `EarlyStopping.mode` flexible for `QSPRModel.fitAttached`.
+- `save_params` argument added to `OptunaOptimization` to save the best hyperparameters to the model (default: `True`).
 
 ## New Features
 - Most unit tests now have a variant that checks whether using a fixed random seed gives reproducible results.